Make `GrabStatus` more generic instead of defining it three types
It's easy enough to change the type generic to use this in all three
places, so we might as well.
